pyproject.toml
[project]
name = "agentforge"
version = "0.1.0"
description = "Crypto arbitrage opportunity detector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-binance>=1.0.19",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "anyio>=4.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0", "pytest-mock>=3.12.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
